(rename) the main export
