+++ htmlcrunch/deno.json
@@ -6,7 +6,7 @@
     "check": "deno fmt --check && deno lint && deno check **/*.ts"
   },
   "exports": {
-    ".": "./index.ts"
+    ".": "./mod.ts"
   },
   "fmt": {
     "exclude": ["**/*.md"]
